+++ src/Symfony/Bundle/SecurityBundle/DependencyInjection/Compiler/AddSecurityVotersPass.php
@@ -11,35 +11,14 @@
 
 namespace Symfony\Bundle\SecurityBundle\DependencyInjection\Compiler;
 
-use Symfony\Component\DependencyInjection\Reference;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
+use Symfony\Component\Security\Core\DependencyInjection\AddSecurityVotersPass as BaseAddSecurityVotersPass;
 
 /**
  * Adds all configured security voters to the access decision manager
  *
  * @author Johannes M. Schmitt <schmittjoh@gmail.com>
+ * @deprecated Deprecated in 2.5, to be removed in 3.0. Use Symfony\Component\Security\Core\DependencyInjection\AddSecurityVotersPass instead.
  */
-class AddSecurityVotersPass implements CompilerPassInterface
+class AddSecurityVotersPass extends BaseAddSecurityVotersPass
 {
-    /**
-     * {@inheritDoc}
-     */
-    public function process(ContainerBuilder $container)
-    {
-        if (!$container->hasDefinition('security.access.decision_manager')) {
-            return;
-        }
-
-        $voters = new \SplPriorityQueue();
-        foreach ($container->findTaggedServiceIds('security.voter') as $id => $attributes) {
-            $priority = isset($attributes[0]['priority']) ? $attributes[0]['priority'] : 0;
-            $voters->insert(new Reference($id), $priority);
-        }
-
-        $voters = iterator_to_array($voters);
-        ksort($voters);
-
-        $container->getDefinition('security.access.decision_manager')->replaceArgument(0, array_values($voters));
-    }
 }