+++ src/Symfony/Component/Security/Core/DependencyInjection/AddSecurityVotersPass.php
@@ -1,0 +1,54 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Security\Core\DependencyInjection;
+
+use Symfony\Component\DependencyInjection\Reference;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
+
+/**
+ * Adds all configured security voters to the access decision manager
+ *
+ * @author Johannes M. Schmitt <schmittjoh@gmail.com>
+ */
+class AddSecurityVotersPass implements CompilerPassInterface
+{
+    private $accessDecisionManagerService;
+    private $voterTag;
+
+    public function __construct($accessDecisionManagerService = 'security.access.decision_manager', $voterTag = 'security.voter')
+    {
+        $this->accessDecisionManagerService = $accessDecisionManagerService;
+        $this->voterTag = $voterTag;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function process(ContainerBuilder $container)
+    {
+        if (!$container->hasDefinition($this->accessDecisionManagerService)) {
+            return;
+        }
+
+        $voters = new \SplPriorityQueue();
+        foreach ($container->findTaggedServiceIds($this->voterTag) as $id => $attributes) {
+            $priority = isset($attributes[0]['priority']) ? $attributes[0]['priority'] : 0;
+            $voters->insert(new Reference($id), $priority);
+        }
+
+        $voters = iterator_to_array($voters);
+        ksort($voters);
+
+        $container->getDefinition($this->accessDecisionManagerService)->replaceArgument(0, array_values($voters));
+    }
+}